fix unit tests; refactor for 'utils' package
