+++ toolium/test/test_config_auto_loader.py
@@ -42,70 +42,42 @@ class TestConfigAutoLoader(unittest.TestCase):
 
     def test_add_attributes_to_page_object(self):
 
-        page_element_list = \
+        page_element = \
             {
                 "Text": {
-                    "Name": "Element1",
+                    "Name": "element1",
                     "Locator-Type": "ID",
                     "Locator-Value": "LOCATOR-VALUE1",
                     "Wait-For-Loaded": True
-                },
-                "InputText": {
-                    "Name": "Element2",
-                    "Locator-Type": "XPATH",
-                    "Locator-Value": "LOCATOR-VALUE2",
-                    "Wait-For-Loaded": False
                 }
             }
 
         page_object = PageObjectAutoloaded()
-        add_attributes_to_page_object(page_element_list, page_object)
+        add_attributes_to_page_object(page_element, page_object)
 
-        assert_true(hasattr(page_object, "element_Element1"))
-        assert_true(hasattr(page_object, "element_Element2"))
-
-        assert_true(hasattr(page_object.element_Element1, "if_wait"))
-        assert_equal(page_object.element_Element1.locator, (By.ID, "LOCATOR-VALUE1"))
-        assert_equal(page_object.element_Element1.if_wait, True)
-
-        assert_true(hasattr(page_object.element_Element2, 'if_wait'))
-        assert_equal(page_object.element_Element2.locator, (By.XPATH, "LOCATOR-VALUE2"))
-        assert_equal(page_object.element_Element2.if_wait, False)
+        assert_true(hasattr(page_object, "element_element1"))
+        assert_true(hasattr(page_object.element_element1, "if_wait"))
+        assert_equal(page_object.element_element1.locator, (By.ID, "LOCATOR-VALUE1"))
+        assert_equal(page_object.element_element1.if_wait, True)
 
     def test_add_attributes_to_page_object_no_wait(self):
 
-        page_element_list = \
-            [
-                {
-                    "Text": {
-                        "Name": "element1",
-                        "Locator-Type": "ID",
-                        "Locator-Value": "LOCATOR-VALUE1",
-                        "Wait-For-Loaded": True
-                    }
-                },
-                {
-                    "InputText": {
-                        "Name": "element2",
-                        "Locator-Type": "XPATH",
-                        "Locator-Value": "LOCATOR-VALUE2"
-                    }
+        page_element = \
+            {
+                "InputText": {
+                    "Name": "element2",
+                    "Locator-Type": "XPATH",
+                    "Locator-Value": "LOCATOR-VALUE2"
                 }
-            ]
+            }
 
         page_object = PageObjectAutoloaded()
-        add_attributes_to_page_object(page_element_list, page_object)
+        add_attributes_to_page_object(page_element, page_object)
 
-        assert_true(hasattr(page_object, "element_element1"))
         assert_true(hasattr(page_object, "element_element2"))
-
-        assert_true(hasattr(page_object.element_Element1, "if_wait"))
-        assert_equal(page_object.element_Element1.locator, (By.ID, "LOCATOR-VALUE1"))
-        assert_equal(page_object.element_Element1.if_wait, True)
-
-        assert_true(hasattr(page_object.element_Element2, 'if_wait'))
-        assert_equal(page_object.element_Element2.locator, (By.XPATH, "LOCATOR-VALUE2"))
-        assert_equal(page_object.element_Element2.if_wait, False)
+        assert_true(hasattr(page_object.element_element2, "if_wait"))
+        assert_equal(page_object.element_element2.locator, (By.XPATH, "LOCATOR-VALUE2"))
+        assert_equal(page_object.element_element2.if_wait, False)
 
     def test_init_page_objects(self):
